Large commit of many components
- hoplite.py largely extended, now with main core functionality, dynamic
import of modules and execution of main calling sequence present.
- Added much of Polynomial Chaos (no conditionals, no Multi-Element yet)
source to project. Organised in such a way that extending from PCE to
PCE-derivatives can be done incrementally (this did not happen in v0.5
and it was, frankly, a mess).
- hoplite.hop moved to lib folder to keep only executable scripts in
project root (to improve readability) and simplify configuration
loading.
- Accessors functions added to systemgraph.py for ids and nodes. This
simplifies the access to certain types of nodes, and the inputs property
has been removed as it was now redundant (and probably buggy).
- Small tweaks and fixes across whole codebase.
- Added stubs for different modules.